refactor: Change backend file organization and CORS
The backend's file structure was changed to be more segmented so that each path has its own file.

The CORS issue we encountered was because the URL for the backend wasn't consistent. To fix this, I made config.js which is where all scripts in the frontend should reference in order to get the backend's URL. For testing purposes, you can change the URL in config.js to be localhost.
