+++ backend/.gitignore
@@ -2,4 +2,4 @@ venv/
 .env
 .venv/
 **/__pycache__
-
+.pytest_cache/
