+++ backend/src/models/Email.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+#Model for email data 
+class Email(BaseModel):
+    email: str
