+++ backend/src/models/Password.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+# Model for password data
+class Password(BaseModel):
+    password: str
