+++ backend/src/models/UserCreate.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+# Create user model
+class UserCreate(BaseModel):
+    name: str
+    email: str
+    password: str
