+++ backend/src/paths/index.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+from fastapi.responses import RedirectResponse
+
+router = APIRouter(
+    prefix = "",
+)
+
+@router.get("/", include_in_schema=False)
+def index():
+    return RedirectResponse("/docs")
