+++ backend/src/paths/login.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter
+from pydantic import BaseModel
+
+router = APIRouter(
+    prefix = "/login",
+    tags = ["Login"]
+)
