+++ backend/test/test_user_registration.py
@@ -1,5 +1,5 @@
 from fastapi.testclient import TestClient
-from app import app  
+from ..src.app import app  
 
 client = TestClient(app)
