+++ frontend/src/App.jsx
@@ -1,5 +1,4 @@
 import React, {useState} from "react";
-// import './App.css';
 import { Routes, Route, BrowserRouter } from 'react-router-dom'
 
 /* Web Pages */
