+++ frontend/src/config.js
@@ -1,0 +1,2 @@
+const apiUrl = "http://localhost:8000";
+export default { apiUrl };
