+++ frontend/vite.config.js
@@ -15,4 +15,7 @@ export default defineConfig({
     // since parsing CSS is slow
     css: true,
   },
+  server: {
+    port: 5173,
+  }
 })
